Setting MAIL_URL environment
