+++ imports/startup/both/useraccounts-configuration.js
@@ -4,8 +4,6 @@ import { Accounts } from 'meteor/accounts-base';
 import { TAPi18n } from 'meteor/tap:i18n';
 import { _ } from 'meteor/underscore';
 import { FlowRouter } from 'meteor/kadira:flow-router';
-import { Memberships } from '/imports/api/memberships/memberships.js';
-import { Communities } from '/imports/api/communities/communities.js';
 import { connectMe } from '/imports/api/memberships/methods.js';
 import { moment } from 'meteor/momentjs:moment';
 
@@ -103,41 +101,6 @@ if (Meteor.isClient) {
 }
 
 if (Meteor.isServer) {
-  Accounts.emailTemplates.siteName = 'Honline';
-  Accounts.emailTemplates.from = 'Honline <noreply@honline.hu>';
-
-  Accounts.emailTemplates.enrollAccount = {
-    subject(user) { return 'Enrollment to Honline'},
-    text(user, url) {
-      const membership = Memberships.findOne({ userEmail: user.emails[0].address })
-      const community = membership.community();
-      const adminEmail = community.admin().userEmail();
-
-      return 'Dear ' + user + ','
-      + '\nYou have been added as a member of community ' + community.name + ' with role: ' + membership.role + '.'
-      + '\nIf you think you have been added by accident, or in fact not want to be part of that community,'
-      + 'please contact the community administrator at ' + adminEmail + ', and ask him to remove you.'
-
-      + '\n\nYou have been also invited to join the condominium management system,'
-      + 'where you can follow the community issues, discuss them and even vote on them.'
-      + 'You can start enjoying all its benefits as soon as you register your account with this email address.'
-
-      + '\n\nThe following link takes you to our simple one click registration:\n'
-      + url + '\n\nThanks.';
-    }
-  };
-
-  Accounts.emailTemplates.verifyEmail = {
-    subject(user) { return 'Verify your email address on Honline'},
-    text(user, url) {
-      return 'Dear ' + user + ','
-      + '\nYou have been registered as a user at honline.hu with this email address. '
-      + 'To confirm your registration please verify your email address within two weeks.'
-      + '\n\nTo verify your account email simply click the link below:\n'
-      + url + '\n\nThanks.';
-    }
-  };
-
-
+  process.env.MAIL_URL = Meteor.settings.mailSender;
   Meteor.setInterval(cleanExpiredEmails, moment.duration(1, 'days').asMilliseconds());
 }
