+++ imports/startup/server/index.js
@@ -4,8 +4,8 @@ import '/imports/api/permissions/config.js';
 // This defines a starting set of data to be loaded if the app is loaded with an empty db.
 import './live-fixtures.js';
 
-// This file configures the Accounts package to define the UI of the reset password email.
-import './reset-password-email.js';
+// This file configures the Accounts package to define the UI of the enrollment / verification / reset password email.
+import './accounts-emailtemplates.js';
 
 // Set up some rate limiting and other important security settings.
 import './security.js';
